+++ pkg/controller/dgraph/models/query/container.go
@@ -18,12 +18,7 @@
 package query
 
 import (
-	"fmt"
-
-	"github.com/vmware/purser/pkg/controller/dgraph/models"
-
 	"github.com/Sirupsen/logrus"
-	"github.com/vmware/purser/pkg/controller/utils"
 )
 
 // RetrieveContainerHierarchy returns hierarchy for a given pod
@@ -51,23 +46,6 @@ func RetrieveContainerMetrics(name string) JSONDataWrapper {
 		logrus.Errorf("wrong type of query for container, empty name is given")
 		return JSONDataWrapper{}
 	}
-	secondsSinceMonthStart := fmt.Sprintf("%f", utils.GetSecondsSince(utils.GetCurrentMonthStartTime()))
-	query := `query {
-		parent(func: has(isContainer)) @filter(eq(name, "` + name + `")) {
-			name
-			type
-			cpu: cpu as cpuRequest
-			memory: memory as memoryRequest
-			st as startTime
-			stSeconds as math(since(st))
-			secondsSinceStart as math(cond(stSeconds > ` + secondsSinceMonthStart + `, ` + secondsSinceMonthStart + `, stSeconds))
-			et as endTime
-			isTerminated as count(endTime)
-			secondsSinceEnd as math(cond(isTerminated == 0, 0.0, since(et)))
-			durationInHours as math(cond(secondsSinceStart > secondsSinceEnd, (secondsSinceStart - secondsSinceEnd) / 3600, 0.0))
-			cpuCost: math(cpu * durationInHours * ` + models.DefaultCPUCostPerCPUPerHour + `)
-			memoryCost: math(memory * durationInHours * ` + models.DefaultMemCostPerGBPerHour + `)
-		}
-	}`
+	query := getQueryForContainerMetrics(name)
 	return getJSONDataFromQuery(query)
 }
